Use plain rsync for backups
`rdiff-backup` is extremely annoying in that if you interrupt the backup, it has
to "regress" all its changes before it allows you to do a new backup. It can
take upwards of an hour to do this, and it won't do any more backups until it's
succeeded in regressing everything. You have no way of knowing this has happened
unless you happen to look at the `cron` logs.

Using `rsync`'s `--link-dest` option, we can do incremental backups into their
own time-stamped folders that will never fail in this way. If a backup gets
interrupted, it doesn't hurt anything since each folder is its own island. A
symlink to the previous latest backup allows us to use it as a source for hard
links to unchanged files.

See
http://www.ibm.com/developerworks/aix/library/au-spunix_rsync/index.html#backup
for an explanation of the technique used here.